+++ website/website/assets/js/map-en.js
@@ -220,7 +220,8 @@ function buildMap(...charts){
 			const params = {
 			  "name": String(selectedReport),
 			  "dataset": String(selectedDataset),
-			  "featureId": String(areas)
+			  "featureId": String(areas),
+			  "includeSvg": true
 			}
 			console.log(params)
 			httpPostAsync(JSON.stringify(params), handleGetQuality);