+++ workers/ohsome_quality_analyst/api/request_models.py
@@ -31,6 +31,7 @@ FidFieldEnum = Enum("FidFieldEnum", {name: name for name in get_fid_fields_api()
 
 
 class BaseRequestModel(pydantic.BaseModel):
+    include_svg: bool = False
     bpolys: Optional[dict]
     dataset: Optional[DatasetEnum]
     feature_id: Optional[str]
@@ -94,6 +95,7 @@ class BaseRequestModel(pydantic.BaseModel):
                     },
                 },
                 "feature_id": {"example": "3"},
+                "include_svg": {"example": False},
             }
         }