+++ workers/tests/integrationtests/test_api_report_geojson_io.py
@@ -117,6 +117,28 @@ class TestApiReportIo(unittest.TestCase):
         content = response.json()
         self.assertEqual(content["type"], "RequestValidationError")
 
+    @oqt_vcr.use_cassette()
+    def test_report_include_svg(self):
+        feature = get_geojson_fixture("heidelberg-altstadt-feature.geojson")
+        url = "/report?name={0}&bpolys={1}&includeSvg={2}".format(
+            self.report_name, feature, True
+        )
+        response = self.client.get(url)
+        result = response.json()
+        self.assertIn("indicators.0.result.svg", list(result["properties"].keys()))
+
+        url = "/report?name={0}&bpolys={1}&includeSvg={2}".format(
+            self.report_name, feature, False
+        )
+        response = self.client.get(url)
+        result = response.json()
+        self.assertNotIn("indicators.0.result.svg", list(result["properties"].keys()))
+
+        url = "/report?name={0}&bpolys={1}".format(self.report_name, feature)
+        response = self.client.get(url)
+        result = response.json()
+        self.assertNotIn("indicators.0.result.svg", list(result["properties"].keys()))
+
 
 if __name__ == "__main__":
     unittest.main()